Move ConfigChangeChecker util to avoid circular imports
