+++ pkg/aggregator/main_test.go
@@ -9,13 +9,14 @@
 package aggregator
 
 import (
-	"github.com/DataDog/datadog-agent/pkg/util/testutil"
 	"os"
 	"testing"
+
+	"github.com/DataDog/datadog-agent/pkg/config"
 )
 
 func TestMain(m *testing.M) {
-	checker := testutil.NewConfigChangeChecker()
+	checker := config.NewChangeChecker()
 	exit := m.Run()
 	if checker.HasChanged() {
 		os.Exit(1)
